Changed Admin to take in command line arg for JSON

+++ admin.py
@@ -1,31 +1,39 @@
 import json
+import sys
 import firebase_admin
 from firebase_admin import db, credentials, firestore
 
-# Connect to firebase
-cred = credentials.Certificate('credentials.json')
-app = firebase_admin.initialize_app(cred)
-db = firestore.client()
-movies = db.collection("movies")
+# Make sure the user passed in a file
+if len(sys.argv) != 2:
+    print("Expected 2 arguments: python admin.py <data.json>")
 
-# Delete all the documents in the collection
-docs = db.collection("movies").list_documents()
-for doc in docs:
-    doc.delete()
+else:
+    try:
+        # Connect to firebase
+        cred = credentials.Certificate('credentials.json')
+        app = firebase_admin.initialize_app(cred)
+        db = firestore.client()
+        movies = db.collection("movies")
 
-# Add movies from json file
-with open('netflix.json') as data_file:
-    data = json.load(data_file)
-    for entry in data:
-        # ------ MAY NEED ------
-        # ----------------------------------------------------------------- 
-        # # Delete entries where the value is "N/A"
-        # to_delete = []
-        # for field in entry:
-        #     if entry[field] == "N/A":
-        #         to_delete.append(field)
-        # for field in to_delete:
-        #     entry.pop(field)
-        # -----------------------------------------------------------------
+        # Delete all the documents in the collection
+        docs = db.collection("movies").list_documents()
+        for doc in docs:
+            doc.delete()
 
-        db.collection("movies").add(entry)
+        # Add movies from json file
+        with open(sys.argv[1]) as data_file:
+            data = json.load(data_file)
+            for entry in data:
+                # ----------------------------------------------------------------- 
+                # Delete entries where the value is "N/A"
+                to_delete = []
+                for field in entry:
+                    if entry[field] == "N/A":
+                        to_delete.append(field)
+                for field in to_delete:
+                    entry.pop(field)
+                # -----------------------------------------------------------------
+
+                db.collection("movies").add(entry)
+    except:
+        print("Error: File not found")
